int=read

+++ tcp/tcp_handle.cc
@@ -61,7 +61,7 @@ void MockEndpoint::OnShutdown(basic::EpollServer* eps, int fd){
 void MockEndpoint::OnReadEvent(int fd){
     char buffer[kBufferSize];
     while(true){
-        size_t nbytes=read(fd,buffer,kBufferSize);  
+        int nbytes=read(fd,buffer,kBufferSize);  
         //https://github.com/eklitzke/epollet/blob/master/poll.c  
         if (nbytes == -1) {
             //if(errno == EWOULDBLOCK|| errno == EAGAIN){}
